Crate keyed_arena

KeyedArena is a bump allocator which allows for random access to its allocations.

Rationale

bumpalo is a popular Rust crate which implements a bump allocator. The function signature for allocation a slice in bumpalo looks something like:

fn alloc_slice_default<'a, T: Default>(&'a self, n: usize) -> &'a [T]

But, now if we want to store the arena alongside references to values allocated into it, we have a problem

struct OurFunStruct {
    arena: bumpalo::Bump,
    things: &'????? [usize], // what lifetime do we want?
}

The “right” solution to this problem would be to have a self-referential type: a type which would expose the lifetime of the arena to be used for its own fields. Unfortunately, this isn’t possible in Rust (except through the use of async/await).

Since we can’t do the “right” thing, we’ll work around it! And that’s the point of KeyedArena

KeyedArena

To avoid the need for self-referential types, KeyedArena allocation outputs a key, which can be used to retrieve the allocation later.

struct OurFunStruct {
    arena: KeyedArena,
    things: AllocationKey<usize>,
}
impl OurFunStruct {
    pub fn new(n: usize) -> Self {
        let arena = KeyedArena::default();
        let things = arena.alloc_slice_fill_with(n, |i| i);
        // Do stuff with things
        let key = things.key();
        std::mem::drop(things);
        Self { arena, things: key }
    }
    pub fn step2(&mut self) -> usize {
        let things = self.arena.borrow_mut(self.things);
        things.iter().copied().sum()
    }
}

Structs

AllocationKey

A key that can be used to retrieve a previous allocation of a slice of type T.

BorrowedAllocation

A mutable allocation of a slice of type T, borrowed from a KeyedArena.

KeyedArena

A bump allocating arena where allocations can be indexed with keys.

KeyedArenaFromPool

A wrapper for a KeyedArena which will return it to a KeyedArenaPool on drop.

KeyedArenaPool

A thread-safe object pool to cache KeyedArena values.

KeyedArenaStats